feat: update ast branch, add support for VM nodes
I had to split up the mutual inductive because it was getting out of
hand, and this last change would have pulled `Command` and everything
it depends on into the giant mutual inductive.

To break the cycle, the rarely used local notation and VM command nodes
are now indirected through an ID, where they are looked up in the
context. This requires some additional complexity in `Parse.M.run`
to tie the knot.