fix: simplify field name heuristic
When mathport has to translate a field name (`f` in `x.f`), the correct result must be already in the FieldNameMap, so the old heuristic did not make a lot of sense. Now we at least use a name that exists.
This is more of an issue since the introduction of "smart naming" which sometimes generates `names_likeThis`.